parser: fix unary number operators parsing
Prior this patch numbers with unary ops were recognized by
the lexer with regex `[+-]?[0-9]*`.

Is leads to incorrect behavior in the cases `int a = +3` vs `int a=b+3`.
In the second case parser expects `+` but got `+3` as a number.

So, the processing of unary operators with numbers are moved to the
parser.

Also, test files were moved to the root of repository to be reused
in other crates.